requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
imutils==0.5.4
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
requests==2.31.0
urllib3==2.0.3
